+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A graph can now be duplicated within the explorer using its context menu or using the shortcut `CTRL+D`. - #291
 - Introduced so-called User Variables which behave like environment variables per graph hierarchy. These can be edited using the `Edit` menu in the graph view. - #294
 - Added drop shadow effect to nodes. - #304
+- Added Undo/Redo commands for pausing a node, renaming a node, and adding dynamic ports. - #322, #315
 - *Internal:* Added support for scope object, that allows executing graphs with a custom source datatree. - #32
 
 ### Changed
@@ -27,7 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed a bug where small mouse movements would not cause the widget to resize. - #288
-- Fixed theme not updating Graph View, Scene, and Nodes correctly. - #301
+- Fixed theme not updating Graph View, Scene, and Nodes c orrectly. - #301
+- Fixed a bug where deleting ports of input/output providers would result in an undefined/unstable internal node state. - #323
+- The graph editor now updates its title when the associated graph changes its object name. - #322
 - *Internal:* `GraphBuilder` now also allows creating connections between ports with different but compatible type ids.
 
 ## [0.14.0] - 2025-04-04